Various fixes on floating-point comparison

+++ source/render/render_hdr.c
@@ -2,22 +2,22 @@
 /*
  *             DO WHAT THE FUCK YOU WANT TO PUBLIC LICENSE
  *                     Version 2, December 2004
- * 
+ *
  *  Copyright (C) 2004 Sam Hocevar <sam@hocevar.net>
- * 
+ *
  *  Everyone is permitted to copy and distribute verbatim or modified
  *  copies of this license document, and changing it is allowed as long
  *  as the name is changed.
  *
  *             DO WHAT THE FUCK YOU WANT TO PUBLIC LICENSE
  *    TERMS AND CONDITIONS FOR COPYING, DISTRIBUTION AND MODIFICATION
- * 
+ *
  *   0. You just DO WHAT THE FUCK YOU WANT TO.
  */
 
 /*
  * HDRState *render_hdr_get_singleton()
- * 
+ *
  * Returns the singleton of the HDR state object.
  * Can be used to query HDR parameters.
  */
@@ -28,11 +28,11 @@ HDRState *render_hdr_get_singleton()
 
 /*
  * void render_hdr_free()
- * 
+ *
  * Frees the HDR state.
- * This struct only contains HDR parameters, and thus it's safe 
- * to be freed after the call to render_hdr() without affecting the 
- * "real" HDR being rendered on the screen, but you won't be able 
+ * This struct only contains HDR parameters, and thus it's safe
+ * to be freed after the call to render_hdr() without affecting the
+ * "real" HDR being rendered on the screen, but you won't be able
  * to query parameters afterwards.
  */
 void render_hdr_free()
@@ -42,9 +42,9 @@ void render_hdr_free()
 
 /*
  * void render_hdr_set_queued( __In BOOL state )
- * 
+ *
  * Queue HDR for rendering.
- * 
+ *
  */
 void render_hdr_set_queued( __In BOOL state )
 {
@@ -53,15 +53,15 @@ void render_hdr_set_queued( __In BOOL state )
 
 /*
  * void render_hdr_new()
- * 
+ *
  * Allocates memory for the HDR state singleton, and set to default values.
- * Overriding values is done through direct query of the singleton (render_hdr_get_singleton()->field), 
+ * Overriding values is done through direct query of the singleton (render_hdr_get_singleton()->field),
  * or through render_hdr_setup(). Neither must be done before HDR is activated with render_hdr().
  */
 void render_hdr_new()
 {
 	HDRState_singleton = (HDRState *) sys_malloc(sizeof(HDRState));
-	
+
 	HDRState_singleton->rt_factor = 4;
 	HDRState_singleton->bit_depth = 32;
 	HDRState_singleton->light_scattering = 0.001;
@@ -71,13 +71,13 @@ void render_hdr_new()
 	HDRState_singleton->highpass_middle_grey = 0.18;
 	HDRState_singleton->highpass_white_cutoff = 0.8;
 	HDRState_singleton->adaption_speed = 2;
-	
+
 	HDRState_singleton->queued = false;
 }
 
 /*
  * BOOL render_hdr_get_queued()
- * 
+ *
  * Yields true if HDR is queued (queried for rendering), false otherwise.
  */
 BOOL render_hdr_get_queued()
@@ -87,7 +87,7 @@ BOOL render_hdr_get_queued()
 
 /*
  * void render_hdr_setup(...)
- * 
+ *
  * Overrides default HDR parameters with a new one.
  * Certain arguments can be ignored; simply pass (-1) to which parameters you want to ignore.
  */
@@ -104,31 +104,31 @@ void render_hdr_setup(
 {
 	if(HDRState_singleton)
 	{
-		HDRState_singleton->rt_factor                   = ifelse(RT == -1, HDRState_singleton->rt_factor, RT);
-		HDRState_singleton->bit_depth                   = ifelse(bit_depth == -1, HDRState_singleton->bit_depth, bit_depth);
-		
-		HDRState_singleton->light_scattering            = ifelse(light_scattering == -1, HDRState_singleton->light_scattering, light_scattering);
-		HDRState_singleton->blur                        = ifelse(blur == -1, HDRState_singleton->blur, blur);
-		HDRState_singleton->bloom_strength              = ifelse(bloom_strength == -1, HDRState_singleton->bloom_strength, bloom_strength);
-		HDRState_singleton->highpass_luminance          = ifelse(highpass_luminance == -1, HDRState_singleton->highpass_luminance, highpass_luminance);
-		HDRState_singleton->highpass_middle_grey        = ifelse(highpass_middle_grey == -1, HDRState_singleton->highpass_middle_grey, highpass_middle_grey);
-		HDRState_singleton->highpass_white_cutoff       = ifelse(highpass_white_cutoff == -1, HDRState_singleton->highpass_white_cutoff, highpass_white_cutoff);
-		HDRState_singleton->adaption_speed              = ifelse(adaption_speed == -1, HDRState_singleton->adaption_speed, adaption_speed);
+		HDRState_singleton->rt_factor                   = ifelse((int) RT == -1, HDRState_singleton->rt_factor, RT);
+		HDRState_singleton->bit_depth                   = ifelse((int) bit_depth == -1, HDRState_singleton->bit_depth, bit_depth);
+
+		HDRState_singleton->light_scattering            = ifelse((int) light_scattering == -1, HDRState_singleton->light_scattering, light_scattering);
+		HDRState_singleton->blur                        = ifelse((int) blur == -1, HDRState_singleton->blur, blur);
+		HDRState_singleton->bloom_strength              = ifelse((int) bloom_strength == -1, HDRState_singleton->bloom_strength, bloom_strength);
+		HDRState_singleton->highpass_luminance          = ifelse((int) highpass_luminance == -1, HDRState_singleton->highpass_luminance, highpass_luminance);
+		HDRState_singleton->highpass_middle_grey        = ifelse((int) highpass_middle_grey == -1, HDRState_singleton->highpass_middle_grey, highpass_middle_grey);
+		HDRState_singleton->highpass_white_cutoff       = ifelse((int) highpass_white_cutoff == -1, HDRState_singleton->highpass_white_cutoff, highpass_white_cutoff);
+		HDRState_singleton->adaption_speed              = ifelse((int) adaption_speed == -1, HDRState_singleton->adaption_speed, adaption_speed);
 	}
 }
 
 /*
  * void render_hdr()
- * 
+ *
  * Performs HDR rendering.
  */
 void render_hdr()
 {
 	__render_hdr_initialize();
-	
+
 	#ifdef __DOF
 	{
-		if(render_dof_get_queued())	
+		if(render_dof_get_queued())
 		    view_dof->stage = view_hdrDownsample;
 		else
 		    camera->stage = view_hdrDownsample;
@@ -136,7 +136,7 @@ void render_hdr()
 	#else
 		camera->stage = view_hdrDownsample;
 	#endif
-	
+
 	view_hdrDownsample->stage = view_hdrHighpass;
 	view_hdrHighpass->stage = view_hdrBlur;
 	view_hdrBlur->stage = view_hdrHBlur;
@@ -147,10 +147,10 @@ void render_hdr()
 __static void __render_hdr_initialize()
 {
 	view_hdrGamma->flags |= (SHOW);
-	
+
 	__static float rt_factor = (render_hdr_get_singleton())->rt_factor;
 	__static float bits = (render_hdr_get_singleton())->bit_depth;
-	
+
 	mtl_hdr->skill4 = floatv(rt_factor);
 	mtl_hdrDownsample->skill1 = floatv(rt_factor);
 	mtl_hdrBlur->skill1 = floatv((render_hdr_get_singleton())->light_scattering);
@@ -162,7 +162,7 @@ __static void __render_hdr_initialize()
 	mtl_hdrHighpass->skill4 = floatv((render_hdr_get_singleton())->highpass_white_cutoff);
 	mtl_hdrGamma3->skill1 = floatv((render_hdr_get_singleton())->adaption_speed);
 	mtl_hdr->skill1 = floatv((render_hdr_get_singleton())->highpass_luminance);
-	
+
 	view_hdrDownsample->arc = camera->arc;
 	view_hdrDownsample->size_x = screen_size.x/rt_factor;
 	view_hdrDownsample->size_y = screen_size.y/rt_factor;
@@ -186,21 +186,21 @@ __static void __render_hdr_initialize()
 	view_hdr->arc = camera->arc;
 	view_hdr->size_x = screen_size.x;
 	view_hdr->size_y = screen_size.y;
-	
+
 	view_hdrGamma->stage = view_hdrGamma2;
 	view_hdrGamma2->stage = view_hdrGamma3;
 	view_hdrGamma3->stage = view_hdrGamma4;
 	view_hdrGamma4->bmap = (RenderState_get_singleton())->bmap_gamma;
-	
+
 	mtl_hdr->skin1 = (RenderState_get_singleton())->map_scene;
-	
+
 	#ifdef __DOF
 		if( (render_dof_get_singleton())->queued)
 		{
 			mtl_hdr->skin1 = RenderState_singleton->bmap_dof;
 		}
 	#endif
-	
+
 	mtl_hdr->skin2 = (RenderState_get_singleton())->bmap_gamma;
 	mtl_hdrHighpass->skin1 = (RenderState_get_singleton())->bmap_gamma;
 	mtl_hdrGamma->skin1 = (RenderState_get_singleton())->map_scene;
